Fix bug causing issues while navigating on facets
