+++ react/package.json
@@ -49,7 +49,7 @@
     "react-dom": "^16.8.3",
     "react-test-renderer": "^16.7.0",
     "regenerator-runtime": "^0.13.1",
-    "typescript": "3.8.3"
+    "typescript": "3.9.7"
   },
   "version": "3.76.0"
 }
